moved products files into pages folder
